number sequence: implement constnat/linear sequence analysis
A sequence of the order zero or one will now be correctly detected and
continued by xlseq. Sequences are assumed to be of an order:
	1) No more than 3
	2) No more than the coresponding Pascal's Triangle which can be
	   drawn by the sample space
In the future, I may change around the sample process to allow for the
entire pattern argument to be passed to the number pattern runner.